fix: Support portless diagrams in custom collector
